refactor(tab): sépare la classe pure de l'enregistrement du tag (mode headless)
