+++ packages/core/src/components/tab-group/tab-group.browser.test.ts
@@ -7,9 +7,9 @@
  */
 import { fixture, html, expect, aTimeout } from '@open-wc/testing';
 import type { ArTabGroup } from './tab-group.js';
-import './tab-group.js';
-import '../tab/tab.js';
-import '../tab-panel/tab-panel.js';
+import './index.js';
+import '../tab/index.js';
+import '../tab-panel/index.js';
 
 function getTab(el: ArTabGroup, panel: string): HTMLElement {
     const tab = el.querySelector<HTMLElement>(`ar-tab[panel="${panel}"]`);
