+++ packages/core/src/components/tab/index.ts
@@ -1,0 +1,11 @@
+import { ArTab } from './tab.js';
+
+customElements.define('ar-tab', ArTab);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-tab': ArTab;
+    }
+}
+
+export { ArTab };
