+++ packages/core/src/index.ts
@@ -34,8 +34,11 @@ export { AnchoredController } from './controllers/anchored.controller.js';
 export { TooltipController } from './controllers/tooltip.controller.js';
 export { ArTooltip } from './components/tooltip/tooltip.js';
 export type { ArTooltipPlacement } from './components/tooltip/tooltip.js';
+import './components/tab-group/index.js';
 export { ArTabGroup } from './components/tab-group/tab-group.js';
+import './components/tab/index.js';
 export { ArTab } from './components/tab/tab.js';
+import './components/tab-panel/index.js';
 export { ArTabPanel } from './components/tab-panel/tab-panel.js';
 export { ArTableSort } from './components/table-sort/table-sort.js';
 export type { TableSortType, TableSortOrder } from './components/table-sort/table-sort.js';
